TINY-10869: Add tests for non-inheritable styles and multiple children
